Backing up partial and reworked code
This was an attempt to integrate VideoJS into the app but, it turned out to be a lot more complicated than originally thought. This was just a test app at first that played one hard-coded identifier. Then, I integrated it into freetv-viewer. This produced problems. So, I create API files on the server to pull in playlists, filter and normalize, and write them to /cache so the frontend (VideoPlayerUI) had clean normalized data. But, the data from Internet Archive is a mess. I pulled all code from freetv-viewer and dumped it here for archival purposes. Much of the code here isn't even referenced in the index.jsx